Refactoring, huge de-bloat
Finally got rid of "storage" and "tabs"\"activeTab" permissions. Remade all the functions to rely on messaging system, so that the context menu would be grayed out when no content scripts responded.
In the same vein, if it's active, that means that all the tags are ready to be fetched - click away!