Hotfix for DA, AS, PX
Reverted the narrowed scopes introduced in 572e00b.

Deviantart, Artstation and Pixiv use AJAX transitions, so, if one happened to start on a main page of the site and then navitage to an individual post via AJAX, SIR fails to parse the new page (scripts are not injected on such transition). While F5 on a target page fixes that, the usability goes down.